refactor: move weight_sampling* to core/data/sampling/
