+++ simba/core/models/ordinal/embedder_multitask.py
@@ -3,8 +3,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from simba.core.data.sampling.weight_sampling import WeightSampling
 from simba.core.models.transformers.embedder import Embedder
-from simba.weight_sampling import WeightSampling
 
 
 class CustomizedCrossEntropyLoss(nn.Module):
